all ids as long and changed vault secret structure

+++ src/FreeSecur.API.Domain/Entities/OrganisationUserRights/OrganisationUserRight.cs
@@ -11,16 +11,16 @@ namespace FreeSecur.API.Domain.Entities.OrganisationUserRights
     {
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public int Id { get; set; }
+        public long Id { get; set; }
         [FsRequired]
-        public int OrganisationUserId { get; set; }
+        public long OrganisationUserId { get; set; }
         [FsRequired]
         public OrganisationUserRightType AccessRight { get; set; }
 
         [FsRequired]
-        public int CreatedById { get; set; }
+        public long CreatedById { get; set; }
         [FsRequired]
-        public int ModifiedById { get; set; }
+        public long ModifiedById { get; set; }
         [FsRequired]
         public DateTime CreatedOn { get; set; }
         [FsRequired]
