+++ src/FreeSecur.API.Domain/Entities/TeamUsers/TeamUser.cs
@@ -13,16 +13,16 @@ namespace FreeSecur.API.Domain.Entities.TeamUsers
     {
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public int Id { get; set; }
+        public long Id { get; set; }
         [FsRequired]
-        public int UserId { get; set; }
+        public long UserId { get; set; }
         [FsRequired]
-        public int TeamId { get; set; }
+        public long TeamId { get; set; }
 
         [FsRequired]
-        public int CreatedById { get; set; }
+        public long CreatedById { get; set; }
         [FsRequired]
-        public int ModifiedById { get; set; }
+        public long ModifiedById { get; set; }
         [FsRequired]
         public DateTime CreatedOn { get; set; }
         [FsRequired]
