+++ src/FreeSecur.API.Domain/Entities/Vaults/Vault.cs
@@ -13,7 +13,7 @@ namespace FreeSecur.API.Domain.Entities.Vaults
     {
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public int Id { get; set; }
+        public long Id { get; set; }
         [FsRequired]
         public string Name { get; set; }
         [FsRequired]
@@ -23,9 +23,9 @@ namespace FreeSecur.API.Domain.Entities.Vaults
         public string MasterKeySalt { get; set; }
 
         [FsRequired]
-        public int CreatedById { get; set; }
+        public long CreatedById { get; set; }
         [FsRequired]
-        public int ModifiedById { get; set; }
+        public long ModifiedById { get; set; }
         [FsRequired]
         public DateTime CreatedOn { get; set; }
         [FsRequired]
